Support overriding charset
Until now, the underlying implementation has assumed UTF-8, and used that
internally when converting byte[] to String (Mysto FPE expects strings).
If, for whatever reason, the user HAS TO use another encoding, using UTF-8
internally had the unfortunate effect of transforming the plaintext, typically resulting
in non-compliant ciphertexts (e.g. with increased size), and no option to go
back to the original plaintext.

By allowing the user to override the charset to apply before dispatching to
Mysto FPE, we can get around this, and at least guarantee for the original and restored plaintexts to be the same, regardess of string encoding.
However, the ciphertexts will be different.

Moral is: USE UTF-8 (!)